refactor(note): move note size constants into their own module
noteWindow.js documents itself as having no knowledge of the store, but
importing the minimums from store.js broke that. Put the default and
minimum note sizes in noteSize.js so store.js, displayUtils.js,
noteWindow.js and main.js share one definition without the window and
placement code depending on the persistence layer.
